refactor: Move retrieval of parameters to input
This patch creates a custom symfony/console `InputInterface`
implementation, `ParamAwareInput`, which decorates an existing
`InputInterface` implementation, and adds the method `getParam()` to it.
The proxy class also wraps over:

- The output
- The current question helper instance
- Parameters from the current command

When `getParam()` is called, it will attempt to retrieve the associated
option, returning the value if found. Otherwise, it will prompt using
the question associated with the parameter, the question helper in the
proxy, and the output from the proxy.

I have updated `InputParamTrait` to add a `run()` method; it calls on
the parent `run()` method, but by first decorating the provided `$input`
in a `ParamAwareInput`.

This approach allows removal of the custom `Application` instance
entirely, and makes consumption of the new input parameters more in line
with options and arguments, in that you now call:

```php
$value = $input->getParam('some-name');
```

instead of

```php
$value = $this->getParam('some-name');
```

as the prior implementation relied heavily on the _application_ being
the correct type.

Essentially, if you can call `$this->addParam()` in your command,  you
can call `$input->getParam()`, and the workflow is completely
self-contained in your command instance.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>