refactor: move original InputParamTrait to new `Console` subnamespace
Since the `InputParamTrait` as originally written targets commands
specifically, I have moved it to a new subnamespace in order to
communicate that more clearly.

In turn, I have renamed `InputParamMethodsTrait` to `InputParamTrait`
within the `Input` namespace, as it helps consumers fulfill the
`InputParamInterface`.

Finally, I renamed `ValidatedQuestionTrait` to `StandardQuestionTrait`.
The trait was not providing a validated question, but rather a standard
question instance that consumers can then modify and return from their
own `getQuestion()` methods.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>