fix: ensure tests pass when prompting multiple times
Due to a bug in symfony/console v5 releases
(https://github.com/symfony/symfony/issues/37046), when specifying
multiple inputs, only the first is used in test environments, which
breaks tests that verify that questions re-prompt on invalid input. The
solution is to set the max attempts to a reasonably high value during
testing. I have accomplished that by setting an enviornment variable,
which I then check for before asking the question.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>