refactor: override run() in LazyLoadingCommand
This makes more sense, as we want to pass on execution to the command
itself. This also ensures that the command's `run()` method is called,
which ensures that, in the case of a `ParamAwareCommand`, we decorate
the input.

`execute()` then becomes essentially a no-op; I've marked it to return
`0` to ensure it fulfills its return typehint.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>