feat: adds ability to mark parameters as multi-value
In symfony/console, you can provide a "mode" to options, which is
generally one of "none" or "required". However, you can also mask
"required" values with a mode indicating the option is an array; in this
scenario, the user can supply the option multiple times, and the
aggregate is presented as an array when the command retrieves the option
value.

Since parameters are essentially options with the ability to prompt for
values if they were not provided, this patch adds support for that mode.

The support is added at the `AbstractParamAwareInput` level, when
calling `getParam()`. The class now has logic to determine if the mode
indicates an array of values, and if so, it will then prompt repeatedly
until an empty line (i.e., user presses "return") is provided. If the
value is not required, the user can press "return" immediately;
otherwise, they will be prompted until at least one value is provided.

The various questions created have been updated to provide information
to the user indicating that multiple values may be entered, and also to
let them know when a value MUST be entered (i.e., parameter is
required).

I have tests for all functionality, which were guided by actual usage
within an application (where I was able to determine what did and did
not work.)

Fixes #23

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>