fix: remove lazy loading functionality
Per #28, the `configure()` method of a command is generally used to
configure a command's metadata: the description, arguments, and options.
In some occasions, those operations require values passed as constructor
arguments.

With the previous LazyLoadingCommand feature added in #12, this could
lead to a few problems:

- Missing command metadata, which would lead to users not getting
  information required to consume the command correctly.
- Warnings interspersed with command information, generally due to
  misconfigured or missing information normally provided via the
  constructor.

This patch removes the lazy loading features. The primary rationale for
them were to prevent scenarios where dependencies that performed
expensive operations during instantiation or `configure()` would lead to
slow command-line usage. These are likely going to be rare, and should
be dealt with on a case-by-case basis by decorating the problematic
commands — instead of _all_ commands.

Fixes #28.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>