qa: tests to validate we do not decorate commands
Modifies the `ExampleDependency` to accept a required value via its
constructor, and provide a method to retrieve it later, using types to
ensure that if the property is not present, the method call will raise a
TypeError.

Modifies the `ExampleCommandWithDependencies` to pull information from
the `ExampleDependency` in its `configure()` method in order to populate
an option's default value, thus reproducing the issue in #28.

Finally, adds tests that validate that text set in `configure()` is
present when calling `list` or `help <command_name>` and using this
command.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>