refactor: remove setOptionMode in favor of setAllowMultipleFlag
This patch removes the method `setOptionMode()` from both the
`InputParamInterface` and `AbstractInputParam` implementation.

It also changes the visiblity of the `$optionMode` property to
`protected`, to allow constructors to change its value.

Finally, it adds a new trait, `AllowMultipleTrait`, defining the method
`setAllowMultipleFlag(bool $flag): self`; based on the value of `$flag`,
it will either add or remove the `InputOption::VALUE_IS_ARRAY` bit from
the option mode.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>