feat: do not mark local commands as third-party commands
This patch provides functionality to exclude local commands (defined as anything not in the vendor directory) from resulting in a prompt warning that a third-party command is about to be executed (when executed as part of a chain).

It does so by by checking to see if it is defined in the configured vendor directory.
The logic excludes classes installed via Laminas, Laminas API Tools, or Mezzio packages.

Fixes #44

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>