qa: fix failing unit test
Recent versions of symfony/console pull the formatter from the output, and inject it with the style formatter; we need to do that as well, but not _require_ it for the test to pass.

Doing so exposed another subtle change, however, in how autocompletion works, as it now emits the input back to the output interface.
This got triggered when exercising the `PathParam`, which required a slight change to expectations to continue to pass.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>