feat: allow providing a custom event dispatcher instance
This patch adds logic to the `ApplicationFactory` to test for a `Laminas\Cli\SymfonyEventDispatcher` instance and, if found, use it to populate the symfony/console instance (if not found, we instantiate one directly, as we did before).

We **require** that the service returned implement `Symfony\Component\EventDispatcher\EventDispatcherInterface` to ensure that we can add our `ConsoleEvents::TERMINATE` event.

Users who want to supply additional listeners or subscribers can do so by defining the service and optionally adding delegators to it.

Fixes #47

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>