qa: fix problem with exit code in LATEST testing envs
For some reason, recent versions of symfony/console allow commands to return a `null` value instead of an int, though they document the return value as an int via docblock.
The result is that not casting the return value to an int prior to calling `setExitCode()` in the `TerminateListener` can lead to an error condition if the value is `null`.
Unfortunately, it also means that Psalm flags a redundant and/or contradiction, requiring a baseline change.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>