fix: ensure normalization of parameters passed as options occurs before validation
When a `Question` is asked, symfony/console performs normalizations prior to validations.
However, in `AbstractParamAwareInput::getOption()`, we were doing the opposite when the option was passed during invocation, which was leading to false negative validations.
This patch now normalizes values passed as options before validating them.

Doing so caused an existing test to fail, however.
The test was validating that an input marked as being an array of values would be marked invalid, with a specific exception message; the exception was now being thrown from an assertion during normalization.
As such, this patch removes the `Assert::isArray` from `normalizeValue()`, and instead applies normalizations without assertions, by either normalizing a non-array value, or applying normalizations to all values in an array.

Fixes #55

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>